src_api/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
paho-mqtt==1.3.1
pytz==2018.4
vine==1.1.4
Werkzeug==0.14.1
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
paho-mqtt==1.3.1
pytz==2018.4
redis==2.10.6
vine==1.1.4
Werkzeug==0.14.1
